Refactor actor.js to typescript and bring part of the item creation logic from the store to the actor class

+++ src/actors/actor.ts
@@ -3,35 +3,29 @@ import LocalisationServer from "../system/localisation_server.js";
 import NotificationServer from "../system/notifications.js";
 import THE_EDGE from "../system/config-the-edge.js";
 
-/**
- * Extend the base Actor document to support attributes and groups with a custom template creation dialog.
- * @extends {Actor}
- */
 export class TheEdgeActor extends Actor {
   async update(data={}, operation={}) {
     this.system.onUpdate(data);
     await super.update(data, operation)
   }
 
+
   /**
    * Is this Actor used as a template for other Actors?
-   * @type {boolean}
    */
-  get isTemplate() {
+  get isTemplate(): boolean {
     return !!this.getFlag("the_edge", "isTemplate");
   }
 
-  interpretCheck(type, roll) {
-    return this.diceServer.interpretCheck(type, roll);
-  }
 
-  get itemWeight() {
+  get itemWeight(): number {
     return this.items.reduce(
       (a, b) => a + ((b.system?.quantity || 1) * b.system?.weight || 0), 0
     );
   }
 
-  learnSkill(newSkill) {
+
+  learnSkill(newSkill: foundryAny): boolean {
     for (const skill of this.items) {
       if (skill.name == newSkill.name && skill.type == newSkill.type && skill.system.level) {
         // Skill already exists and can potentially be leveled
@@ -41,9 +35,11 @@ export class TheEdgeActor extends Actor {
     }
 
     // Skill doesn't exist yet
-    if (!this.fulfillsRequirements(newSkill, true)) return false;
-    const ph = this.system.PracticeHours
-    let cost = Aux.getSkillCost(newSkill)
+    if (!this.fulfillsRequirements(newSkill, 1)) return false;
+    const ph = this.system.PracticeHours;
+    const cost = Aux.getSkillCost(newSkill, "increase")
+    if (typeof cost === "undefined") return false;
+
     if (cost <= ph.max - ph.used) {
       this.update({"system.PracticeHours.used": ph.used + cost})
       return true;
@@ -58,9 +54,10 @@ export class TheEdgeActor extends Actor {
     return false
   }
 
-  fulfillsRequirements(skill, newSkill = false) {
+
+  fulfillsRequirements(skill: foundryAny, newSkillLevel: number = 0) {
     if (skill.type == "Languageskill") return true;
-    const level = skill.system.level - newSkill;
+    const level = skill.system.level - newSkillLevel;
     const requirements = skill.system.requirements[level];
     if (requirements === undefined || requirements.length == 0) return true;
 
@@ -68,7 +65,7 @@ export class TheEdgeActor extends Actor {
       const group = requirement.group;
       const details = structuredClone(requirement);
       if (group == "skills") {
-        const skillRef = this.items.filter(x => 
+        const skillRef = this.items.filter((x: Item) => 
           x.name.toLowerCase() == requirement.field.toLowerCase());
         if (skillRef.length == 0) {
           NotificationServer.notify("Missing requirements", details)
@@ -91,10 +88,26 @@ export class TheEdgeActor extends Actor {
     return true;
   }
 
-  skillLevelIncrease(skillID) {
+
+  async addOneItem(item: Item): Promise<void> {
+    const existingCopy = this.findItem(item);
+    if (existingCopy && "quantity" in item.system) {
+      await existingCopy.update({"system.quantity": existingCopy.system.quantity + 1});
+    } else {
+      const itemCls = getDocumentClass("Item");
+      const newSystem = {...item.system};
+      newSystem.quantity = 1;
+      await itemCls.create({name: item.name, type: item.type, system: newSystem}, {parent: this});
+    }
+  }
+
+
+  skillLevelIncrease(skillID: string): boolean {
     let skill = this.items.get(skillID)
     if (!this.fulfillsRequirements(skill)) return false;
-    let cost = Aux.getSkillCost(skill, "increase");
+    const cost = Aux.getSkillCost(skill, "increase");
+    if (typeof cost === 'undefined') return false;
+
     const ph = this.system.PracticeHours
     if (cost < ph.max - ph.used) {
       this.update({"system.PracticeHours.used": ph.used + cost})
@@ -104,43 +117,46 @@ export class TheEdgeActor extends Actor {
     return false
   }
 
-  skillLevelDecrease(skillID) {
-    let skill = this.items.get(skillID)
-    let level = skill.system.level
-    let gain = Aux.getSkillCost(skill, "decrease")
+
+  skillLevelDecrease(skillID: string): void {
+    const skill = this.items.get(skillID);
+    const level = skill.system.level;
+    const gain = Aux.getSkillCost(skill, "decrease");
+    if (typeof gain === 'undefined') return;
+
     const ph = this.system.PracticeHours
     this.update({"system.PracticeHours.used": ph.used - gain})
     if (level > 1) skill.update({"system.level": level - 1});
-    else skill.delete()
+    else skill.delete();
   }
 
-  deleteSkill(skillID) {
-    const skill = this.items.get(skillID)
-    let gain = Aux.getSkillCost(skill, "delete");
+
+  deleteSkill(skillID: string): void {
+    const skill = this.items.get(skillID);
+    const gain = Aux.getSkillCost(skill, "delete");
+    if (typeof gain === 'undefined') return;
+    
     const ph = this.system.PracticeHours
     this.update({"system.PracticeHours.used": ph.used - gain})
-    skill.delete()
+    skill.delete();
   }
 
-  async addOrCreateVantage(vantage) {
-    let AP = this.system.AdvantagePoints;
+
+  async addOrCreateVantage(vantage: Item): Promise<void> {
+    const AP = this.system.AdvantagePoints;
 
     // Can be created or leveled?
     if (vantage.type == "Advantage" && vantage.system.AP + AP.used > AP.max) {
-      let msg = LocalisationServer.parsedLocalisation(
-        "AP missing", "Notifications",
+      NotificationServer.notify(
+        "AP missing",
         {name: vantage.name, need: vantage.system.AP, available: AP.max - AP.used}
-      )
-      ui.notifications.notify(msg);
-      return false;
+      );
+      return;
     } 
-    const existingCopy = this.findItem(vantage)
+    const existingCopy = this.findItem(vantage);
     if (existingCopy && existingCopy.system.level >= existingCopy.system.maxLevel) {
-      let msg = LocalisationServer.parsedLocalisation(
-        "Max Level", "Notifications", {name: vantage.name}
-      )
-      ui.notifications.notify(msg)
-      return false;
+      NotificationServer.notify("Max Level", {name: vantage.name})
+      return;
     } 
 
     // Now create or level
@@ -150,26 +166,26 @@ export class TheEdgeActor extends Actor {
 
     if (existingCopy) {
       const sys = existingCopy.system
-      if (sys.maxLevel > sys.level) {
-        await existingCopy.update({"system.level": sys.level + 1})
-      }
+      await existingCopy.update({"system.level": sys.level + 1})
     } else {
       const cls = getDocumentClass("Item");
-      const newVantage = await cls.create({name: vantage.name, type: vantage.type}, {parent: this});
-      newVantage.update({"system": vantage.system});
+      await cls.create(
+        {name: vantage.name, type: vantage.type, system: {...vantage.system, level: 1}},
+        {parent: this}
+      );
     }
     await this.update(update);
   }
 
-  decrementVantage(vantage) {
+
+  decrementVantage(vantage: Item): void {
     const AP = this.system.AdvantagePoints;
     const itemAP = vantage.system.AP;
     if (vantage.type == "Disadvantage" && AP.max - itemAP < AP.used) {
-      let msg = LocalisationServer.parsedLocalisation(
-        "AP missing decrement", "Notifications", {name: vantage.name, need: itemAP, available: AP.max - AP.used}
-      )
-      ui.notifications.notify(msg);
-      return undefined;
+      NotificationServer.notify(
+        "AP missing decrement", {name: vantage.name, need: itemAP, available: AP.max - AP.used}
+      );
+      return;
     }
 
     if (vantage.type == "Advantage") this.update({"system.AdvantagePoints.used": AP.used - itemAP});
@@ -179,15 +195,15 @@ export class TheEdgeActor extends Actor {
     else vantage.delete();
   }
 
-  deleteVantage(vantage) {
+
+  deleteVantage(vantage: Item): void {
     const AP = this.system.AdvantagePoints;
     const itemAP = (vantage.system.hasLevels ? vantage.system.level : 1) * vantage.system.AP;
     if (vantage.type == "Disadvantage" && AP.max - itemAP < AP.used) {
-      let msg = LocalisationServer.parsedLocalisation(
-        "AP missing deletion", "Notifications", {name: vantage.name, need: itemAP, available: AP.max - AP.used}
+      NotificationServer.notify(
+        "AP missing deletion", {name: vantage.name, need: itemAP, available: AP.max - AP.used}
       )
-      ui.notifications.notify(msg);
-      return undefined;
+      return;
     }
 
     if (vantage.type == "Advantage") this.update({"system.AdvantagePoints.used": AP.used - itemAP});
@@ -195,26 +211,28 @@ export class TheEdgeActor extends Actor {
     vantage.delete();
   }
 
-  findItem(item) {
-    let _existingCopy = false
+
+  findItem(item: Item): Item | undefined {
+    let existingCopy: Item | undefined = undefined;
     for (const _item of this.itemTypes[item.type]) {
       if (_item.name == item.name) {
         if (_item.type == "Ammunition") {
           const _cap = _item.system.capacity;
           const cap = item.system.capacity;
           if (_cap.max == cap.max && _cap.value == cap.value && !_item.system.loaded) {
-            _existingCopy = _item
+            existingCopy = _item
           }
         } else {
-          _existingCopy = _item
+          existingCopy = _item
         }
       }
     }
-    return _existingCopy;
+    return existingCopy;
   }
 
-  getSkillEffects(onlyActive = false) {
-    function skillFilter(skill) {
+
+  getSkillEffects(onlyActive: boolean = false): Record<string, any>[] {
+    function skillFilter(skill: Item): boolean {
       const hasEffect = skill.system.modifiers.length > 0;
       const isActive = !onlyActive || skill.system.active;
       return hasEffect && isActive;
@@ -233,8 +251,9 @@ export class TheEdgeActor extends Actor {
     });
   }
 
-  getItemEffects(onlyEquipped = false) {
-    function itemFilter(item) {
+
+  getItemEffects(onlyEquipped: boolean = false): Record<string, any>[] {
+    function itemFilter(item: Item): boolean {
       const hasEffect = item.system.modifiers.length > 0;
       const isEquipped = !onlyEquipped || item.system.equipped;
       return hasEffect && isEquipped;
@@ -252,21 +271,26 @@ export class TheEdgeActor extends Actor {
     });
   }
 
-  attachOuterArmour(armourId, shellId, tokenId) {
+
+  attachOuterArmour(armourId: string, shellId: string, tokenId: string) {
     const armour = this.items.get(armourId);
     const shell = this.items.get(shellId);
     const availableAttachment = armour.system.attachmentPoints.max - armour.system.attachmentPoints.used;
     if (shell.system.attachmentPoints.max > availableAttachment) {
-      let msg = LocalisationServer.parsedLocalisation(
-        "Missing Attachment points", "Notifications",
+      NotificationServer.notify(
+        "Missing Attachment points",
         {available: availableAttachment, needed: shell.system.attachmentPoints.max}
       )
-      ui.notifications.notify(msg)
+      return;
     }
+
     // Hack relevant information into the shells attachment list, needed in item.js upon breaking
-    shell.update({"system.equipped": true, "system.attachments": [{actorId: this.id, tokenId: tokenId, armourId: armour._id}]});
+    shell.update({
+      "system.equipped": true,
+      "system.attachments": [{actorId: this.id, tokenId: tokenId, armourId: armour.id}]
+    });
     const attachments = armour.system.attachments;
-    attachments.push({actorId: this.id, tokenId: tokenId, shellId: shell._id, shell: shell});
+    attachments.push({actorId: this.id, tokenId: tokenId, shellId: shell.id, shell: shell});
     armour.update({
       "system.attachments": attachments,
       "system.attachmentPoints.used": armour.system.attachmentPoints.used + shell.system.attachmentPoints.max
